Move the utils dir so that it is not run like a test dir
