+++ cypress/utils/room-utils.ts
@@ -1,17 +1,17 @@
 import Chainable = Cypress.Chainable;
-import TchapCreateRoom from '../../../src/lib/createTchapRoom';
-import { TchapRoomType } from '../../../src/@types/tchap';
+import TchapCreateRoom from '../../src/lib/createTchapRoom';
+import { TchapRoomType } from '../../src/@types/tchap';
 export default class RoomUtils {
-    static createPublicRoom(roomName: string): Chainable<string> {
+    public static createPublicRoom(roomName: string): Chainable<string> {
         return cy.createRoom(TchapCreateRoom.roomCreateOptions(roomName, TchapRoomType.Forum, false).createOpts);
     }
-    static createPrivateRoom(roomName: string): Chainable<string> {
+    public static createPrivateRoom(roomName: string): Chainable<string> {
         return cy.createRoom(TchapCreateRoom.roomCreateOptions(roomName, TchapRoomType.Private, false).createOpts);
     }
-    static createPrivateWithExternalRoom(roomName: string): Chainable<string> {
+    public static createPrivateWithExternalRoom(roomName: string): Chainable<string> {
         return cy.createRoom(TchapCreateRoom.roomCreateOptions(roomName, TchapRoomType.External, false).createOpts);
     }
-    static openRoomAccessSettings(roomName: string) {
+    public static openRoomAccessSettings(roomName: string) {
         //open room
         cy.get('[aria-label="'+roomName+'"]').click();
         cy.get('.mx_RoomHeader_chevron').click();
